refactor : model -> entity 패키지 명 변경
