+++ src/main/java/com/app/guttokback/group/domain/repository/GroupMemberRepository.java
@@ -1,8 +1,8 @@
 package com.app.guttokback.group.domain.repository;
 
-import com.app.guttokback.group.domain.model.GroupMember;
-import com.app.guttokback.group.domain.model.SubscriptionGroup;
-import com.app.guttokback.user.domain.model.User;
+import com.app.guttokback.group.domain.entity.GroupMember;
+import com.app.guttokback.group.domain.entity.SubscriptionGroup;
+import com.app.guttokback.user.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GroupMemberRepository extends JpaRepository<GroupMember, Long> {
