+++ src/main/java/com/app/guttokback/group/domain/repository/SubscriptionGroupRepository.java
@@ -1,6 +1,6 @@
 package com.app.guttokback.group.domain.repository;
 
-import com.app.guttokback.group.domain.model.SubscriptionGroup;
+import com.app.guttokback.group.domain.entity.SubscriptionGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SubscriptionGroupRepository extends JpaRepository<SubscriptionGroup, Long> {
