+++ src/main/java/com/app/guttokback/notification/application/service/NotificationService.java
@@ -3,15 +3,15 @@ package com.app.guttokback.notification.application.service;
 import com.app.guttokback.common.api.PageOption;
 import com.app.guttokback.common.api.PageResponse;
 import com.app.guttokback.notification.application.dto.NotificationListResponse;
+import com.app.guttokback.notification.domain.entity.Notification;
 import com.app.guttokback.notification.domain.enums.Category;
 import com.app.guttokback.notification.domain.enums.Message;
 import com.app.guttokback.notification.domain.enums.Status;
-import com.app.guttokback.notification.domain.model.Notification;
 import com.app.guttokback.notification.domain.repository.NotificationQueryRepository;
 import com.app.guttokback.notification.domain.repository.NotificationRepository;
-import com.app.guttokback.user.domain.model.User;
+import com.app.guttokback.user.domain.entity.User;
+import com.app.guttokback.userSubscription.domain.entity.UserSubscription;
 import com.app.guttokback.userSubscription.domain.enums.Subscription;
-import com.app.guttokback.userSubscription.domain.model.UserSubscription;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
