+++ src/main/java/com/app/guttokback/userSubscription/application/service/UserSubscriptionService.java
@@ -6,15 +6,15 @@ import com.app.guttokback.common.exception.CustomApplicationException;
 import com.app.guttokback.common.exception.ErrorCode;
 import com.app.guttokback.email.application.service.ReminderService;
 import com.app.guttokback.user.application.service.UserService;
-import com.app.guttokback.user.domain.model.User;
+import com.app.guttokback.user.domain.entity.User;
 import com.app.guttokback.userSubscription.application.dto.controllerDto.response.UserSubscriptionListResponse;
 import com.app.guttokback.userSubscription.application.dto.serviceDto.SubscriptionListInfo;
 import com.app.guttokback.userSubscription.application.dto.serviceDto.SubscriptionSearchInfo;
 import com.app.guttokback.userSubscription.application.dto.serviceDto.UserSubscriptionSaveInfo;
 import com.app.guttokback.userSubscription.application.dto.serviceDto.UserSubscriptionUpdateInfo;
+import com.app.guttokback.userSubscription.domain.entity.UserSubscription;
 import com.app.guttokback.userSubscription.domain.enums.PaymentCycle;
 import com.app.guttokback.userSubscription.domain.enums.Subscription;
-import com.app.guttokback.userSubscription.domain.model.UserSubscription;
 import com.app.guttokback.userSubscription.domain.repository.UserSubscriptionQueryRepository;
 import com.app.guttokback.userSubscription.domain.repository.UserSubscriptionRepository;
 import lombok.RequiredArgsConstructor;
