+++ src/main/java/com/app/guttokback/userSubscription/domain/repository/UserSubscriptionQueryRepository.java
@@ -1,7 +1,7 @@
 package com.app.guttokback.userSubscription.domain.repository;
 
 import com.app.guttokback.common.api.PageOption;
-import com.app.guttokback.userSubscription.domain.model.UserSubscription;
+import com.app.guttokback.userSubscription.domain.entity.UserSubscription;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Repository;
 import java.time.LocalDate;
 import java.util.List;
 
-import static com.app.guttokback.user.domain.model.QUser.user;
-import static com.app.guttokback.userSubscription.domain.model.QUserSubscription.userSubscription;
+import static com.app.guttokback.user.domain.entity.QUser.user;
+import static com.app.guttokback.userSubscription.domain.entity.QUserSubscription.userSubscription;
 
 @Repository
 @RequiredArgsConstructor
