+++ src/main/java/com/app/guttokback/userSubscription/domain/repository/UserSubscriptionRepository.java
@@ -1,6 +1,6 @@
 package com.app.guttokback.userSubscription.domain.repository;
 
-import com.app.guttokback.userSubscription.domain.model.UserSubscription;
+import com.app.guttokback.userSubscription.domain.entity.UserSubscription;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
