+++ src/test/java/com/app/guttokback/email/application/service/EmailTemplateServiceTest.java
@@ -1,12 +1,12 @@
 package com.app.guttokback.email.application.service;
 
 import com.app.guttokback.email.application.dto.serviceDto.EmailInfo;
-import com.app.guttokback.user.domain.model.User;
+import com.app.guttokback.user.domain.entity.User;
 import com.app.guttokback.user.domain.repository.UserRepository;
+import com.app.guttokback.userSubscription.domain.entity.UserSubscription;
 import com.app.guttokback.userSubscription.domain.enums.PaymentCycle;
 import com.app.guttokback.userSubscription.domain.enums.PaymentMethod;
 import com.app.guttokback.userSubscription.domain.enums.Subscription;
-import com.app.guttokback.userSubscription.domain.model.UserSubscription;
 import com.app.guttokback.userSubscription.domain.repository.UserSubscriptionRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
