+++ src/test/java/com/app/guttokback/group/application/service/SubscriptionGroupServiceTest.java
@@ -2,10 +2,10 @@ package com.app.guttokback.group.application.service;
 
 import com.app.guttokback.common.exception.CustomApplicationException;
 import com.app.guttokback.group.application.dto.serviceDto.SubscriptionGroupSaveInfo;
-import com.app.guttokback.group.domain.model.SubscriptionGroup;
+import com.app.guttokback.group.domain.entity.SubscriptionGroup;
 import com.app.guttokback.group.domain.repository.GroupMemberRepository;
 import com.app.guttokback.group.domain.repository.SubscriptionGroupRepository;
-import com.app.guttokback.user.domain.model.User;
+import com.app.guttokback.user.domain.entity.User;
 import com.app.guttokback.user.domain.repository.UserRepository;
 import com.app.guttokback.userSubscription.domain.enums.PaymentCycle;
 import com.app.guttokback.userSubscription.domain.enums.PaymentMethod;
