+++ src/test/java/com/app/guttokback/group/domain/repository/GroupMemberRepositoryTest.java
@@ -1,8 +1,8 @@
 package com.app.guttokback.group.domain.repository;
 
-import com.app.guttokback.group.domain.model.GroupMember;
-import com.app.guttokback.group.domain.model.SubscriptionGroup;
-import com.app.guttokback.user.domain.model.User;
+import com.app.guttokback.group.domain.entity.GroupMember;
+import com.app.guttokback.group.domain.entity.SubscriptionGroup;
+import com.app.guttokback.user.domain.entity.User;
 import com.app.guttokback.user.domain.repository.UserRepository;
 import com.app.guttokback.userSubscription.domain.enums.PaymentCycle;
 import com.app.guttokback.userSubscription.domain.enums.PaymentMethod;
