+++ src/test/java/com/app/guttokback/subscription/application/service/UserSubscriptionServiceTest.java
@@ -3,7 +3,7 @@ package com.app.guttokback.subscription.application.service;
 import com.app.guttokback.common.api.PageOption;
 import com.app.guttokback.common.api.PageResponse;
 import com.app.guttokback.common.exception.CustomApplicationException;
-import com.app.guttokback.user.domain.model.User;
+import com.app.guttokback.user.domain.entity.User;
 import com.app.guttokback.user.domain.repository.UserRepository;
 import com.app.guttokback.userSubscription.application.dto.controllerDto.response.UserSubscriptionListResponse;
 import com.app.guttokback.userSubscription.application.dto.serviceDto.SubscriptionListInfo;
@@ -11,11 +11,11 @@ import com.app.guttokback.userSubscription.application.dto.serviceDto.Subscripti
 import com.app.guttokback.userSubscription.application.dto.serviceDto.UserSubscriptionSaveInfo;
 import com.app.guttokback.userSubscription.application.dto.serviceDto.UserSubscriptionUpdateInfo;
 import com.app.guttokback.userSubscription.application.service.UserSubscriptionService;
+import com.app.guttokback.userSubscription.domain.entity.UserSubscription;
 import com.app.guttokback.userSubscription.domain.enums.PaymentCycle;
 import com.app.guttokback.userSubscription.domain.enums.PaymentMethod;
 import com.app.guttokback.userSubscription.domain.enums.PaymentStatus;
 import com.app.guttokback.userSubscription.domain.enums.Subscription;
-import com.app.guttokback.userSubscription.domain.model.UserSubscription;
 import com.app.guttokback.userSubscription.domain.repository.UserSubscriptionRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
